// src/app/models/tipoHabitaciones/tipoHabitaciones-response.ts
export class TipoHabitacionesResponse {
    idTipoHabitacion: number;
    descripcion: string;
    precio: number;
    capacidad: number;
    imagenUrl: string;
}

export class TipoHabitaciones {
    idTipoHabitacion: number;
    descripcion: string;
    precio: number;
    capacidad: number;
    imagenUrl: string;
}